+++ front-office/src/config.ts
@@ -1,2 +1,2 @@
 // export const HOST_IP = import.meta.env.VITE_HOST_IP;
-export const HOST_IP = "172.25.224.1";
+export const HOST_IP = "192.168.16.130";
